==> Code/0. Experiment - Getting Data.ipynb <==
#import libraries 

import os
import requests

BASE_DIR = '..'
NIGHTLIGHTS_DIR = os.path.join(BASE_DIR, 'data/Nightlights')

# 2013 is the closest year the NOAA maintains
for year in [2013]:
    year = str(year)
    url = 'https://ngdc.noaa.gov/eog/data/web_data/v4composites/F18' + year + '.v4.tar'
    target = os.path.join(NIGHTLIGHTS_DIR, year)
    os.makedirs(target, exist_ok=True)
    target += '/nightlights_data'
    response = requests.get(url, stream=True)
    if response.status_code == 200:
        with open(target, 'wb') as f:
            f.write(response.raw.read())
            print(f'Saved to {target}')

import tarfile
import gzip
import shutil

# this unzips the data
for year in [2013]:
    print(year)
    year = str(year)
    folder_loc = os.path.join(NIGHTLIGHTS_DIR, year)
    file_loc = folder_loc + '/nightlights_data'
    tar = tarfile.open(file_loc)
    tar.extractall(path=folder_loc)
    
    files = os.listdir(os.path.join(NIGHTLIGHTS_DIR, year))
    for file in files:
        path = os.path.join(NIGHTLIGHTS_DIR, year, file)
        if 'stable' in file: # only need stable_lights
            if file.split('.')[-1] == 'gz':
                # unzip the file is a .gz file
                with gzip.open(path, 'rb') as f_in:
                    with open(path[:-3], 'wb') as f_out:
                        shutil.copyfileobj(f_in, f_out)
                        print(f'Saved to {path}')
        os.remove(path)

pwd()
