==> Code/1. OSM Prep - Extract Buildings Features.ipynb <==
### OpenStreetMap Data Preprocessing - Extract Buildings Feature

#Mount google drive for content synchronization

from google.colab import drive
drive.mount('/content/drive')


#Install libraries for geospatial data wrangling and analysis

#Geopandas for wrangling
!pip install geopandas

#Rtree
!!apt install python3-rtree

#Contextily for basemap plotting 
!apt-get install libproj-dev proj-data proj-bin
!apt-get install libgeos-dev
!pip install cython
!pip install cartopy
!pip3 install contextily==1.0rc2


### Step 1: Load OSM data for Buildings

import geopandas
import pandas as pd
import numpy as np
from geopandas import GeoDataFrame
import geopandas
from shapely.geometry import LineString, Point
import matplotlib.pyplot as plt
import contextily

building_df = geopandas.read_file('/content/drive/My Drive/Thesis/Bangladesh_OSM_Data/gis_osm_buildings_a_free_1.shp')
building_df.tail()

#Load cluster data
clusters_df = pd.read_csv('/content/drive/My Drive/Thesis/DHS_Data/bangladesh_cluster_avg_asset_2013_updated_new.csv')
clusters_df = geopandas.GeoDataFrame(clusters_df, geometry = geopandas.points_from_xy(clusters_df.longitude, clusters_df.latitude))
clusters_df.tail()

OSM Building Data Feature Engineering 

#Build new dataframe the rows of which has info on types of building

building_df = building_df[building_df['type'].notna()]
building_df.tail()

building_df.head()

#Separate Urban and Rural area clusters

clusters_rural = clusters_df[clusters_df['URBAN_RURA'] == 'R']
clusters_urban = clusters_df[clusters_df['URBAN_RURA'] == 'U']
clusters_rural.head()

clusters_urban.head()

type(clusters_rural.geometry[0])

#convert Rural clusters to buffers with 5km radius
clusters_rural.crs = {'init' :'epsg:4326'}
clusters_rural= clusters_rural.to_crs(epsg=3174)  
buffer_length_in_meters = (5 * 1000)

rural_buffer = clusters_rural.geometry.buffer(buffer_length_in_meters)

len(rural_buffer)

#convert Urban clusters to buffers with 2km radius
clusters_urban.crs = {'init' :'epsg:4326'}
clusters_urban= clusters_urban.to_crs(epsg=3174)  
buffer_length = (2 * 1000)

urban_buffer = clusters_urban.geometry.buffer(buffer_length)

len(urban_buffer)

#Convert roads_df to same crs as Urban and Rural clusters df
building_df = building_df.to_crs(epsg=3174)

len(building_df)

#Helper function to calculate the area of building type within a buffer zone of a cluster

def building_within(buffer,building_df): 
#  building_count_list = []
  building_area_total_list = []
  building_area_mean_list = []
  building_area_pct_list = []
  for point in buffer:
    building_area = building_df.intersection(point).area
    building_area_total = building_area.sum() / (1000**2)
    building_area_mean = building_area.mean() / (1000**2)
    building_area_pct = building_area_total / point.area 
    building_area_total_list.append(building_area_total)
    building_area_mean_list.append(building_area_mean)
    building_area_pct_list.append(building_area_pct)
  return building_area_total_list, building_area_mean_list, building_area_pct_list

#Helper function to calculate the number of building type within a buffer zone of a cluster

def building_count(buffer,building_df): 
  building_count_total = []
  for point in buffer:
    building_number = building_df.within(point).sum()
    building_count_total.append(building_number)
  return building_count_total

#Create list of different POIS types

building_type_name = building_df['type'].unique()
building_type_name

#Create list of different POIS types

building_type_name = building_df['type'].unique()
building_type_name

#Binning different values into one


building_df["type"].replace({"apartments": "residential", "dormatory": "residential", 'apartments': 'residential',\
                             'house': 'residential', 'dormitory': 'residential', 'apartment block': 'residential',\
                             'residential,shop,sch': 'residential', 'residential;commerci': 'residential','resedential': 'residential',
                             'hassan house': 'residential','VC Residence': 'residential', "commercial;residenti": 'residential', 'commercial;residenti': 'residential',
                             'apartment': 'residential','rrdidential':'residential', 'house 2': 'residential','three houses':'residential',
                             'Building_with_grocer':'residential', 'Building_with_laundr':'residential', 'House': 'residential',


                             'mosque': 'religious', 'cathedral': 'religious', 'Chayanot,_cultural_c': 'religious',\
                             'church':'religious', 'temple': 'religious', 'Mosque': 'religious',
                             

                             "Building_with_grocer": "commercial", 'cafe': 'commercial',\
                             'office': 'commercial', 'bank': 'commercial', 'hotel': 'commercial', 'retail': 'commercial',\
                             'warehouse': 'commercial', 'supermarket': 'commercial', 'store': 'commercial', 'Laboratory': 'commercial',
                             'Guesthouse_&_Confere': 'commercial', 'Hostel': 'commercial','11 Stored':'commercial', 'club':'commercial', 'shop':'commercial',
                             'market':'commercial', 'Shop': 'commercial', 'Sports_Building': 'commercial', 'CLC_Power_Company_Li': 'commercial',


                             'Stadium': 'sport', 'grandstand': 'sport','stadium': 'sport','arena': 'sport',


                             'silo': 'industrial', 'factory': 'industrial', 'storage_tank': 'industrial','manufacture': 'industrial',


                             'construction': 'construction', 'under_construction': 'construction', 'Temporary': 'construction',
                             'Semi permanent': 'construction', 'Under Construction':'construction', 'Underconstraction': 'construction',
                             'under construction': 'construction','non-permanent': 'construction', 'under_construction_1':'construction', 'semi-permanent':'construction',
                             'underconstruction': 'construction', 'Semi_permanent': 'construction', 'Semi_permanennt':'construction', 'Permanent':'construction', 'temporary':'construction', 'Semni_permanent': 'construction',

                             
                             "hospital": "health", 'medical': 'health', 'health_clinic': 'health', 'Clinic': 'health',


                             "train_station": "transport service", 'toll_booth': 'transport service', 'gasometer': 'transport service',\
                             'hangar': 'transport','WatchTower': 'transport', 'garage': 'transport', 'garages': 'transport',\
                             'transportation': 'transport', 


                             'university': 'education', 'BUET Admin':'education', 'Admin': 'education', 'school': 'education',\
                             'me workshop': 'education', 'canteen': 'education', 'English_Language_Cen': 'education',\
                             'College': 'education', 'college': 'education','kindergarten': 'education', 'Madrasha': 'education',
                             'madrasa': 'education', 'educational_institut': 'education', 'Recreational': 'education', 'Cantin': 'education',
                             'Nawab_Faizunnesa_Hal': 'education','Madrasa': 'education', 'School': 'education', 'Nawab_faizunnesa_Hal': 'education', 'Gopalpur High School': 'education', 'Institution': 'education',
                             'academic': 'education',

                             'public':'government', 'service': 'government', 'Union Council': 'government', 'Community_Center': 'government',\
                             'Official':'government', 'Goverment Office': 'government', 'Police_Station': 'government', 'goverment': 'government',
                             'infrastructure': 'government','civil_service': 'government','Police_station': 'government', 'civic': 'government',


                             "Hut": 'low-cost', 'Orphanage': 'low-cost', 'shed': 'low-cost', 'cabin': 'low-cost',\
                              'Tin shet': 'low-cost', 'Tin-shade': 'low-cost', 'tin shade house':'low-cost', 'Tin Shed House':'low-cost', 'Tin shaded':'low-cost', 'tin shaded':'low-cost',
                              'hut': 'low-cost', 'Brick_Kiln': 'low-cost', 'Low Income Settlemen': 'low-cost', 'Slum': 'low-cost',
                             'kiosk': 'low-cost', 'Tin Shed':"low-cost",'slum': 'low-cost','Garbage Area':'low-cost', 'Tinshed':'low-cost','stable': 'low-cost',
                             'Tin_shed':'low-cost', 'tin_shed':'low-cost','Tinshed building':'low-cost', 'tinshed': 'low-cost', 'hamlet': 'low-cost','ruins': 'low-cost', 'bungalow': 'low-cost',


                             'Agriculture': 'agriculture', 'barn': 'agriculture', 'farm': 'agriculture', 'farm_auxiliary': 'agriculture', 'greenhouse?':'agriculture',
                             'greenhouse': 'agriculture', 'cowshed': 'agriculture',


                             'mixed': 'other', 'no': 'other', 'roof': 'other', 'PAka': 'other', 'Katcha': 'other', 'new': 'other', 'detached': 'other', 'yes4': 'other', 'unclassified': 'other', 'Cart': 'other',
                             'mix': 'other', 'mixed use': 'other', 'godown': 'other', 'car': 'other','Multipurpose': 'other', 'No':'other', 'no house':'other',
                             'unused':'other', 'sculpture':'other', 'pond':'other', 'risesn':'other', 'wood':'other', 'terrace':'other', 'place':'other', 'uy':'other', 'Marker':'other', 'well':'other', 
                             '3': 'other', 'q': 'other','\\': 'other', 'ola_del_mar':'other','yesS':'other','toilet':'other', 'CO': 'other', 'C': 'other', 'yesqq': 'other', 'yeb': 'other', 'B-49': 'other','333': 'other','yesq': 'other', 'Need a plan': 'other'
                        
                             }, inplace=True)


building_type_name = building_df['type'].unique()
building_type_name

building_df["type"].replace({'Canteen': 'education', 'auditorium': 'sport', 'gallery': 'sport', 'camp': 'sport', 'vc': 'other',
                             'qk ext': 'other', 'post office': 'government', 'Gov._Office': 'government', 'hostel': 'commercial', 
                             'residential,commerci': 'residential', 'water': 'other', 'residential_cum_comm': 'residential', 
                             'Semi-Pucca': 'low-cost', 'shed;residential': 'low-cost', 'Residential_cum_Comm': 'residential',
                             'land': 'other', 'residential, Commerc': 'residential', 'building': 'other', 'residential_&_commer': 'residential', 
                             'rrsidential': 'residential', 'residean': 'residential', 'Medical center': 'health', 'residence': 'residential', 
                             'Yes': 'other', 'Kashipur, Barisal': 'other'}, inplace=True)

building_type_name = building_df['type'].unique()
building_type_name

#Create list of dataframes of different building types
building_type = []
for i in range(len(building_type_name)): 
  building_type_list = building_df[building_df['type']== building_type_name[i]]
  building_type.append(building_type_list) 

##Engineer the total area of building feature for each types and append to rural dataframe

for i in range(len(building_type)):
  building_area_total_list, building_area_mean_list, building_area_pct_list = building_within(rural_buffer, building_df)
  building_area_total_list = np.asarray(building_area_total_list)
  building_area_mean_list = np.asarray(building_area_mean_list)
  building_area_pct_list = np.asarray(building_area_pct_list)
  clusters_rural[f"{building_type_name[i]} building area total"] = building_area_total_list
  clusters_rural[f"{building_type_name[i]} building area mean"] = building_area_mean_list
  clusters_rural[f"{building_type_name[i]} building area proportion"] = building_area_pct_list  

##Engineer the total area of building feature for each types and append to urban dataframe

for i in range(len(building_type)):
  building_area_total_list, building_area_mean_list, building_area_pct_list = building_within(urban_buffer, building_df)
  building_area_total_list = np.asarray(building_area_total_list)
  building_area_mean_list = np.asarray(building_area_mean_list)
  building_area_pct_list = np.asarray(building_area_pct_list)
  clusters_urban[f"{building_type_name[i]} building area total"] = building_area_total_list
  clusters_urban[f"{building_type_name[i]} building area mean"] = building_area_mean_list
  clusters_urban[f"{building_type_name[i]} building area proportion"] = building_area_pct_list  

##Engineer the total area of building feature for each types and append to rural dataframe

#get counts: https://stackoverflow.com/questions/54127731/how-can-i-count-the-number-of-polygons-a-shape-intersects

import pandas as pd
import geopandas as gp
from shapely.geometry import Polygon
from shapely.geometry import Point
import matplotlib.pyplot as plt

rural_buffer_copy = rural_buffer.copy()

# generate spatial index
sindex = building_df.sindex
# define empty list for results
results_list = []
# iterate over the points
for buffer in rural_buffer_copy: 
    # find approximate matches with r-tree, then precise matches from those approximate ones
    possible_matches_index = list(sindex.intersection(buffer.bounds))
    possible_matches = building_df.iloc[possible_matches_index]
    precise_matches = possible_matches[possible_matches.intersects(buffer)]
    results_list.append(len(precise_matches))

# add list of results as a new column
rural_buffer_copy['polygons'] = pd.Series(results_list)

# for i in range(len(building_type)):
#   building_count_total = building_count(rural_buffer, building_df)
#   building_count_total = np.asarray(building_count_total)
#   clusters_rural[f"Count of {building_type_name[i]} building"] = building_count_total


building_df_copy = building_df.copy()
#building_df_copy = gp.GeoDataFrame(building_df_copy, geometry = 'geometry')

sindex = building_df_copy.sindex
print(sindex)

##Engineer the total area of building feature for each types and append to rural dataframe



for i in range(len(building_type)):
  building_count_total = building_count(urban_buffer, building_df)
  building_count_total = np.asarray(building_count_total)
  clusters_urban[f"Count of {building_type_name[i]} building"] = building_count_total


# New Section

import os
import pickle
os.chdir('/content/drive/My Drive/Thesis/OSM_data')
with open('clusters_rural_building', 'wb') as fp:
    pickle.dump(clusters_rural, fp)

import os
import pickle
os.chdir('/content/drive/My Drive/Thesis/OSM_data')
with open('clusters_urban_building', 'wb') as fp:
    pickle.dump(clusters_urban, fp)

with open('clusters_rural_building', 'rb') as fp:
  clusters_rural = pickle.load(fp)

with open('clusters_urban_building', 'rb') as fp:
  clusters_urban = pickle.load(fp)

# New Section

### Merge engineered features with the original data


merged = pd.concat([clusters_rural, clusters_urban])
merged = merged.sort_values(by = 'cluster')

with open('merged_building', 'wb') as fp:
    pickle.dump(merged, fp)

merged.to_csv('clusters_building.csv')
